Add more functionality to Image.  Move solve_field to images.py
