+++ pocs/utils/images/calculations.py
@@ -31,121 +31,6 @@ from .metadata import get_wcsinfo
 quantity_support()
 
 
-def solve_field(fname, timeout=15, solve_opts=[], **kwargs):
-    """ Plate solves an image.
-
-    Args:
-        fname(str, required):       Filename to solve in either .cr2 or .fits extension.
-        timeout(int, optional):     Timeout for the solve-field command, defaults to 60 seconds.
-        solve_opts(list, optional): List of options for solve-field.
-        verbose(bool, optional):    Show output, defaults to False.
-    """
-    verbose = kwargs.get('verbose', False)
-    if verbose:
-        print("Entering solve_field")
-
-    solve_field_script = "{}/scripts/solve_field.sh".format(os.getenv('POCS'), '/var/panoptes/POCS')
-
-    if not os.path.exists(solve_field_script):
-        raise error.InvalidSystemCommand("Can't find solve-field: {}".format(solve_field_script))
-
-    # Add the options for solving the field
-    if solve_opts:
-        options = solve_opts
-    else:
-        options = [
-            '--guess-scale',
-            '--cpulimit', str(timeout),
-            '--no-verify',
-            '--no-plots',
-            '--crpix-center',
-            '--downsample', '4',
-        ]
-        if kwargs.get('clobber', True):
-            options.append('--overwrite')
-        if kwargs.get('skip_solved', True):
-            options.append('--skip-solved')
-
-        if 'ra' in kwargs:
-            options.append('--ra')
-            options.append(str(kwargs.get('ra')))
-        if 'dec' in kwargs:
-            options.append('--dec')
-            options.append(str(kwargs.get('dec')))
-        if 'radius' in kwargs:
-            options.append('--radius')
-            options.append(str(kwargs.get('radius')))
-
-        if os.getenv('PANTEMP'):
-            options.append('--temp-dir')
-            options.append(os.getenv('PANTEMP'))
-
-    cmd = [solve_field_script, ' '.join(options), fname]
-    if verbose:
-        print("Cmd: ", cmd)
-
-    try:
-        proc = subprocess.Popen(cmd, universal_newlines=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    except OSError as e:
-        raise error.InvalidCommand("Can't send command to solve_field.sh. {} \t {}".format(e, cmd))
-    except ValueError as e:
-        raise error.InvalidCommand("Bad parameters to solve_field.sh. {} \t {}".format(e, cmd))
-    except Exception as e:
-        raise error.PanError("Timeout on plate solving: {}".format(e))
-
-    return proc
-
-
-def get_solve_field(fname, **kwargs):
-    """ Convenience function to wait for `solve_field` to finish.
-
-    This function merely passes the `fname` of the image to be solved along to `solve_field`,
-    which returns a subprocess.Popen object. This function then waits for that command
-    to complete, populates a dictonary with the EXIF informaiton and returns. This is often
-    more useful than the raw `solve_field` function
-
-    Parameters
-    ----------
-    fname : {str}
-        Name of file to be solved, either a FITS or CR2
-    **kwargs : {dict}
-        Options to pass to `solve_field`
-
-    Returns
-    -------
-    dict
-        Keyword information from the solved field
-    """
-
-    verbose = kwargs.get('verbose', False)
-    if verbose:
-        print("Entering get_solve_field")
-
-    proc = solve_field(fname, **kwargs)
-    try:
-        output, errs = proc.communicate(timeout=kwargs.get('timeout', 30))
-    except subprocess.TimeoutExpired:
-        proc.kill()
-        output, errs = proc.communicate()
-
-    out_dict = {}
-
-    if errs is not None:
-        warn("Error in solving: {}".format(errs))
-    else:
-        # Read the EXIF information from the CR2
-        if fname.endswith('cr2'):
-            out_dict.update(read_exif(fname))
-            fname = fname.replace('cr2', 'new')  # astrometry.net default extension
-            out_dict['solved_fits_file'] = fname
-
-        try:
-            out_dict.update(fits.getheader(fname))
-        except OSError:
-            if verbose:
-                print("Can't read fits header for {}".format(fname))
-
-    return out_dict
 
 
 def solve_offset(first_dict, second_dict, verbose=False):  # unused
